Some fixes to the markdown-processor
Updated most of the information to Markdown 1.7 and moved the code up the
preprocessor chain to become a TextPreprocessor in order to avoid unnecessary
paragraphs in the final HTML output.

The extra '\n' -> '<br />' replacement is necessary to also allow newlines
in the sourcecode without splitting the output HTML into multiple code
segments.