Improve Rust lifetime highlighting.
A lifetime token isn’t normally a label, though it can be in a couple of
situations, demonstrated thus::

    'foo: loop {
        break 'foo;
    }

Therefore, I’ve changed the highlighting for all other places where lifetimes
appear from ``Name`` or ``Name.Label`` to ``Name.Attribute``. As for the
special lifetime ``'static``, which was being highlighted as ``Keyword``, I
changed it to ``Name.Builtin`` which is a little more accurate.