autopygmentize: update for pygments 2.0
Update the dates, and remove the copyright assertion, which doesn't make
sense for a file in the public domain.

Teach autopygmentize about application/javascript.

Update the list of MIME types that can't be mapped relative to file 5.20 and
pygments 2.0.

Fix a FIXME that needed a 2.0 feature. In the process, remove some
hard-wired settings (-f terminal256 -O style=native) that were not commented
on in the commit that introduced them (changeset 2846).