Implement guessing and resolution between TypeScript and TypoScript.
The new TypoScript lexer was taking precedence over TypeScript, which share
the same file extension. This meant that any previously-working TypeScript
source files would be misidentified and highlighted incorrectly. This was due
to the lack of a guesser for TypeScript, and TypoScript having a naturally
higher priority.

This change balances things a bit between the lexers. TypoScript now has an
explicitly-higher priority than TypeScript, both for the file extension and
defaults when guessing content, as it has less we can rely on in terms of
content. TypeScript has a lower priority, but has an analyse_text() method
that checks for common symbols in code, giving it a better chance of matching
for actual TypeScript source files.

TypeScript also gains an additional file extension, *.tsx. Ideally, this would
have its own lexer that understood inline JSX syntax, but giving general
TypeScript syntax highlighting for these files is a step in the right
direction.