Remove the fallback values from the new analyse_text changes.
These values would trigger problems when guessing content without a filename,
and really aren't needed. They've been removed, allowing the standard behavior
to take place.

On top of this, the regex for TypeScript has been slightly altered to include
a whitespace match after "export" to be within the group for the "export"
text.